drivers: video: Add support for self-buffered, DCMI snapshot mode
The stm32 DCMI driver currently only supports double-buffered continuous
mode, which limits the utility of this driver to only work with
processors that have enough memory to have multiple buffers. The ability
to use a DCMI device in snapshot mode is supported by hardware, and is a
common use case for embedded devices. This adds the ability to operate
DCMI in snapshot mode, and to supply a single buffer to the driver.

Signed-off-by: Isaac Dekine <isaac.dekine@gmail.com>